Se crea la subclase Lavadora, con sus metodos y atributos correspondientes

+++ src/SeventeenthChallenge/Consumo.java
@@ -2,16 +2,14 @@ package SeventeenthChallenge;
 //blancos, negro, rojo, azul y gris
 public enum Consumo {
         A(100), B(80), C(60), D(50), E(30), F(10);
-        private final double consumo;
+        private final int consumo;
 
-        Consumo(double consumo) {
+        Consumo(int consumo) {
                 this.consumo = consumo;
         }
 
-        public double getConsumo() {
+        public int getConsumo() {
                 return consumo;
         }
-
-
 }
